+++ src/components/code_editor/code_editor.js
@@ -126,6 +126,8 @@ export class EuiCodeEditor extends Component {
       ...rest
     } = this.props;
 
+    const dataTestSubj = this.props['data-test-subj'];
+
     const classes = classNames('euiCodeEditorWrapper', {
       'euiCodeEditorWrapper-isEditing': this.state.isEditing,
     });
@@ -199,7 +201,7 @@ export class EuiCodeEditor extends Component {
       <div
         className={classes}
         style={{ width, height }}
-        data-test-subj="codeEditorContainer">
+        data-test-subj={dataTestSubj || 'codeEditorContainer'}>
         {prompt}
 
         <AceEditor
@@ -231,6 +233,7 @@ EuiCodeEditor.propTypes = {
   isReadOnly: PropTypes.bool,
   setOptions: PropTypes.object,
   cursorStart: PropTypes.number,
+  'data-test-subj': PropTypes.string,
 
   /**
    * Use string for a built-in mode or object for a custom mode